ask_web_youtube/requirements.txt
streamlit==1.45.0
pytest==8.3.5
pytest-asyncio==0.26.0
pydocstyle==6.3.0
groq==0.24.0
isodate==0.7.2
httpx==0.27.2
Jinja2==3.1.5
python-dotenv==1.0.1
google-api-python-client==2.166.0
youtube-transcript-api==1.0.3